refactor(Popup): create first internal component popup used by tooltip and popover
