+++ packages/ui/src/internalComponents/Popup/helpers.ts
@@ -1,6 +1,6 @@
 import type { RefObject } from 'react'
 
-export type TooltipPlacement = 'top' | 'right' | 'bottom' | 'left' | 'auto'
+export type PopupPlacement = 'top' | 'right' | 'bottom' | 'left' | 'auto'
 export const ARROW_WIDTH = 8 // in px
 const SPACE = 4 // in px
 const TOTAL_USED_SPACE = ARROW_WIDTH + SPACE // in px
@@ -50,7 +50,7 @@ const computePlacement = ({
 }
 
 type ComputePositionsTypes = {
-  placement: TooltipPlacement
+  placement: PopupPlacement
   childrenRef: RefObject<HTMLDivElement>
   tooltipRef: RefObject<HTMLDivElement>
 }
