+++ packages/ui/src/internalComponents/index.ts
@@ -1,0 +1,1 @@
+export { Popup } from './Popup'
